examples/postgresql/requirements.txt
psycopg2-binary==2.9.10
faker==37.4.0
python-dotenv==1.0.0

# Requirements for sample Semantic RAG application
chromadb>=1.0.12
psycopg2-binary>=2.9.0
requests>=2.31.0
PyYAML>=6.0
python-dotenv>=1.0.0 